+++ components/Input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shallot-ui/component-input
 
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-utils@2.0.0-beta.9
+
 ## 3.0.0-beta.8
 
 ### Patch Changes